fix: fixes after review
