--- .nvmrc
@@ -1,1 +1,0 @@
-v22.18.0
