--- src/components/pages/customers/about/index.tsx
@@ -1,3 +1,0 @@
-import About from "./about";
-
-export default About
