+++ src/components/pages/customers/hero.tsx
@@ -157,7 +157,9 @@ function CustomerCard({
               : `${ROUTE.customers}/${slug.current}`
           }
           className="group absolute top-0 left-0 z-10 h-full w-full"
-        />
+        >
+          <span className="sr-only">visit customer</span>
+        </Link>
 
         {config.background && (
           <>
@@ -192,9 +194,9 @@ function CustomerCard({
           height={logo.asset.metadata.dimensions.height || 40}
           priority
         />
-        <h3 className="relative mt-auto text-xl leading-snug font-normal tracking-tighter md:max-w-[530px] md:text-2xl">
+        <h1 className="relative mt-auto text-xl leading-snug font-normal tracking-tighter md:max-w-[530px] md:text-2xl">
           “{title}”
-        </h3>
+        </h1>
         <p className="relative mt-2 text-sm leading-snug font-[350] tracking-tight text-gray-8 md:text-base">{`${author} — ${author_position}`}</p>
         <span className="hidden-start relative mt-4 flex items-center gap-x-1.5 text-lagune-3 transition-colors duration-200 group-hover:text-lagune-2">
           {link.type === "story" ? "Read story" : "Visit site"}
