--- src/components/pages/customers/quote/index.ts
@@ -1,3 +1,0 @@
-import Quote from "./quote";
-
-export default Quote
