--- src/components/pages/customers/related/index.ts
@@ -1,3 +1,0 @@
-import Related from "./related";
-
-export default Related
