--- src/components/pages/customers/reviews/index.ts
@@ -1,3 +1,0 @@
-import Reviews from "./reviews";
-
-export default Reviews
