--- src/components/pages/customers/socials/index.ts
@@ -1,3 +1,0 @@
-import Socials from "./socials";
-
-export default Socials
