+++ src/components/shared/slider/index.ts
