+++ src/styles/slider.css
@@ -37,3 +37,5 @@
     }
   }
 }
+
+// Other sections
